omg, it runs, not sure if it works yet though
