+++ optima/parameters.py
@@ -7,7 +7,8 @@ Version: 2016jan28
 """
 
 from numpy import array, isnan, zeros, argmax, mean, log, polyfit, exp, maximum, minimum, Inf, linspace, median, shape
-from optima import OptimaException, odict, printv, sanitize, uuid, today, getdate, smoothinterp, dcp, defaultrepr, objrepr, getresults, convertlimits
+from optima import OptimaException, odict, printv, sanitize, uuid, today, getdate, smoothinterp, dcp, defaultrepr, objrepr # Utilities 
+from optima import getresults, convertlimits, gettvecdt # Heftier functions
 
 eps = 1e-3 # TODO WARNING KLUDGY avoid divide-by-zero
 
@@ -506,33 +507,14 @@ def makesimpars(pars, inds=None, keys=None, start=2000, end=2030, dt=0.2, tvec=N
     for key in keys: # Loop over all keys
         if issubclass(type(pars[key]), Par): # Check that it is actually a parameter -- it could be the popkeys odict, for example
             try: simpars[key] = pars[key].interp(tvec=simpars['tvec'], dt=dt, smoothness=smoothness) # WARNING, want different smoothness for ART
-            except: raise OptimaException('Could not figure out how to interpolate parameter "%s"' % key)
+            except OptimaException as E: 
+                errormsg = 'Could not figure out how to interpolate parameter "%s"' % key
+                errormsg += 'Error: "%s"' % E.message
+                raise OptimaException(errormsg)
 
     return simpars
 
 
-def gettvecdt(tvec=None, dt=None, justdt=False):
-    ''' 
-    Function to encapsulate the logic of returning sensible tvec and dt based on flexible input.
-    
-    If tvec and dt are both supplied, do nothing.
-    
-    Will always work if tvec is not None, but will use default value for dt if dt==None and len(tvec)==1.
-    
-    Version: 2016jan30
-    '''
-    defaultdt = 0.2 # WARNING, slightly dangerous to hard-code but should be ok, since very rare situation
-    if tvec is None: 
-        if justdt: return defaultdt # If it's a constant, maybe don' need a time vector, and just return dt
-        else: raise OptimaException('No time vector supplied, unable to interpolate') # Usual case, crash
-    elif isinstance(tvec, (int, float)): tvec = array([tvec]) # Convert to 1-element array
-    elif shape(tvec): # Make sure it has a length -- if so, overwrite dt
-        if len(tvec)>=2: dt = tvec[1]-tvec[0] # Even if dt supplied, recalculate it from the time vector
-        elif dt is None: dt = defaultdt # Or give up and use default
-        else: dt = dt # Use input
-    else:
-        raise OptimaException('Could not understand tvec of type "%s"' % type(tvec))
-    return tvec, dt
 
 
 
@@ -556,16 +538,14 @@ def applylimits(y, limits=None, dt=None, warn=True, verbose=2):
     limits = convertlimits(limits=limits, dt=dt, verbose=verbose)
     
     # Apply limits, preserving original class
-    origclass = y.__class__ # To restore original type
     if isinstance(y, (int, float)):
-        newy = origclass(median([limits[0], y, limits[1]]))
+        newy = median([limits[0], y, limits[1]])
         if warn and newy!=y: printv('Warning, parameter value reset from %f to %f' % (y, newy), 4, verbose)
         return newy
     elif shape(y):
-        newy = array(y)
+        newy = array(y) # Make sure it's an array and not a list
         newy[newy<limits[0]] = limits[0]
         newy[newy>limits[1]] = limits[1]
-        newy = origclass(newy)
         if warn and (sum(newy<limits[0]) or sum(newy<limits[0])):
             printv('Warning, parameter value reset from:\n%f\nto:\n%f' % (y, newy), 4, verbose)
     else:
